# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
# during dev, should point to your local machine
host: localhost:8080
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
...
paths:
...
  /api/registration/validateId:
   x-swagger-router-controller: Register
   post:
     tags:
      - "Register"
     operationId: registerUser
     description: User register service
     parameters:
      - name: udid
        in: header
        description: User's udid
        required: true
        type: string
      - name: registerObject
        in: body
        required: true
        schema:
         $ref: "#/definitions/RegisterObject"
     responses:
       "200":
         description: OK
         schema:
          $ref: "#/definitions/RegisterResponse"
  /api/registration/categories:
    x-swagger-router-controller: Register
    get:
      tags:
      - "Register"
      description: Return
      operationId: getEventsCategories
      parameters:
        - name: udid
          in: header
          description: User's udid
          required: true
          type: string
        - name: Authorization
          in: header
          description: User's authorization
          required: false
          type: string
      responses:
        "200":
          description: response
          schema:
            type: array
            items:
              $ref: "#/definitions/Interests"        
...
definitions:
...
  RegisterResponse:
   required:
    - code
    - message
   properties:
    code:
     type: number
    message:
     type: string
    secret:
     type: string
  RegisterObject:
    properties:
     sex:
      enum: ["MALE","FEMALE"]
      type: string 
     mail:
      type: string
     age:
      type: string
     job:
      type: string
     adress:
      type: string
     interests:
      type: array
      items:
       $ref: "#/definitions/Interests"
  Interests:
   properties:
    id:
     type: string
    name:
     type: string
...